refactor: move cityCard to components
