+++ src/components/CityCard/CityCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/constants/main.scss';
+@import '../../assets/styles/constants/main.scss';
 
 .city-card {
   box-sizing: border-box;
